+++ gecko_client/admin_api.go
@@ -10,7 +10,7 @@ const (
 )
 
 type AdminApi struct {
-	rpcRequester geckoJsonRpcRequester
+	rpcRequester jsonRpcRequester
 }
 
 type NodeID struct {
@@ -42,9 +42,8 @@ type GetNodeIDResponse struct {
 	Id int	`json:"id"`
 }
 
-// TODO Maybe parse the response into IPAddr:Port for the user?
 func (api AdminApi) GetPeers() ([]Peer, error) {
-	responseBodyBytes, err := api.rpcRequester.makeRpcRequest(adminEndpoint, "admin.peers")
+	responseBodyBytes, err := api.rpcRequester.makeRpcRequest(adminEndpoint, "admin.peers", make(map[string]interface{}))
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "Error making request")
 	}
@@ -58,7 +57,7 @@ func (api AdminApi) GetPeers() ([]Peer, error) {
 }
 
 func (api AdminApi) GetNodeId() (string, error) {
-	responseBodyBytes, err := api.rpcRequester.makeRpcRequest(adminEndpoint, "admin.getNodeID")
+	responseBodyBytes, err := api.rpcRequester.makeRpcRequest(adminEndpoint, "admin.getNodeID", make(map[string]interface{}))
 	if err != nil {
 		return "", stacktrace.Propagate(err, "Error making request")
 	}