+++ gecko_client/health_api.go
@@ -29,11 +29,11 @@ type GetLivenessResponse struct {
 }
 
 type HealthApi struct {
-	rpcRequester geckoJsonRpcRequester
+	rpcRequester jsonRpcRequester
 }
 
 func (api HealthApi) GetLiveness() (LivenessInfo, error) {
-	responseBodyBytes, err := api.rpcRequester.makeRpcRequest(healthApiEndpoint, "health.getLiveness")
+	responseBodyBytes, err := api.rpcRequester.makeRpcRequest(healthApiEndpoint, "health.getLiveness", make(map[string]interface{}))
 	if err != nil {
 		return LivenessInfo{}, stacktrace.Propagate(err, "Error getting liveness")
 	}